v0.13
Changes:
* Changed Font
* Added Randomized Costume Button
* Added Settings Panel
* Added Sound Effects and Music files
* Added Ending UI

+++ Electric_Manta_Prototype/Assets/Scripts/Save/SaveFile.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class SaveFile
 {
-    public int Coins;
+    public double Coins;
     public int Criminals;
     public int Costume;
